move flashy logic to extension from lsp client
