+++ typescript/vscode-ext/packages/language-server/src/server.ts
@@ -33,7 +33,7 @@ import { exec } from 'child_process'
 // import { FileChangeType } from 'vscode'
 import fs from 'fs'
 // import { cliBuild, cliCheckForUpdates, cliVersion } from './baml-cli'
-import { type ParserDatabase, TestRequest } from '@baml/common'
+// import { type ParserDatabase, TestRequest } from '@baml/common'
 import debounce from 'lodash/debounce'
 // import { TextDocumentIdentifier } from 'vscode-languageserver-protocol'
 import { TextDocument } from 'vscode-languageserver-textdocument'
